FastAPI/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
ultralytics
pillow
sqlmodel
shutil